pipeline works for video sample
